+++ tests/ui/specs/TabLayout/VerticalTabsWithoutIcons/VerticalTabsWithoutIcons-specs.js
@@ -1,62 +1,59 @@
 const Page = require('../TabLayoutPage');
 
 describe('TabLayout', function () {
-
-	beforeEach(function () {
-		Page.open('VerticalTabsWithoutIcons');
+	beforeEach(async function () {
+		await Page.open('VerticalTabsWithoutIcons');
 	});
 
 	describe('vertical tabs without icons', function () {
-
 		describe('collapsing/expanding behavior', function () {
 			describe('5-way interaction', function () {
-
-				it('should collapse tabs when focus is moved to a Spottable component in the content container - [QWT-2756]', function () {
+				it('should collapse tabs when focus is moved to a Spottable component in the content container - [QWT-2756]', async function () {
 					// 5-way down to second tab
-					Page.spotlightDown();
-					Page.tabLayout.view(2).waitForExist();
+					await Page.spotlightDown();
+					await (await Page.tabLayout.view(2)).waitForExist();
 					// Step 3-2: 5-way right to Spottable component
-					Page.waitTransitionEnd(1500, 'waiting for Panel transition', () => {
-						Page.spotlightRight();
+					await Page.waitTransitionEnd(1500, 'waiting for Panel transition', async () => {
+						await Page.spotlightRight();
 					});
 					// check that layout is collapsed
-					expect(Page.tabLayout.isCollapsed).to.be.true();
+					expect(await Page.tabLayout.isCollapsed).to.be.true();
 				});
 
-				it('should expand tabs when focus is moved to a Spottable component in the tabs container - [QWT-2756]', function () {
+				it('should expand tabs when focus is moved to a Spottable component in the tabs container - [QWT-2756]', async function () {
 					// 5-way down to second tab
-					Page.spotlightDown();
-					Page.tabLayout.view(2).waitForExist();
+					await Page.spotlightDown();
+					await (await Page.tabLayout.view(2)).waitForExist();
 					// focus the contents
-					Page.waitTransitionEnd(1500, 'waiting for Panel transition', () => {
-						Page.spotlightRight();
+					await Page.waitTransitionEnd(1500, 'waiting for Panel transition', async () => {
+						await Page.spotlightRight();
 					});
-					expect(Page.tabLayout.isCollapsed).to.be.true();
+					expect(await Page.tabLayout.isCollapsed).to.be.true();
 					// Step 4: Back to the tabs
-					Page.waitTransitionEnd(1500, 'waiting for Panel transition', () => {
-						Page.spotlightLeft();
+					await Page.waitTransitionEnd(1500, 'waiting for Panel transition', async () => {
+						await Page.spotlightLeft();
 					});
-					expect(Page.tabLayout.isCollapsed).to.be.false();
-					expect(Page.tabLayout.tabItems[1].isFocused()).to.be.true();
+					expect(await Page.tabLayout.isCollapsed).to.be.false();
+					expect(await (await Page.tabLayout.tabItems())[1].isFocused()).to.be.true();
 				});
 			});
 			describe('pointer interaction', function () {
-
-				it('should collapse and expand tabs when focus is moved between `Spottable` components in the content and tabs containers via pointer move - [QWT-2757]', function () {
+				it('should collapse and expand tabs when focus is moved between `Spottable` components in the content and tabs containers via pointer move - [QWT-2757]', async function () {
 					// focus the layout's tabs
-					Page.tabLayout.hoverTabs();
+					await Page.tabLayout.hoverTabs();
 					// select the second tab
-					Page.tabLayout.tabItems[1].click();
-					Page.tabLayout.view(2).waitForExist();
+					// console.log(await Page.tabLayout.tabItems);
+					await (await Page.tabLayout.tabItems())[1].click();
+					await (await Page.tabLayout.view(2)).waitForExist();
 
 					// move pointer to Spottable component in content container
-					$('#button2').moveTo();
+					await $('#button2').moveTo();
 					// check that layout is collapsed
-					expect(Page.tabLayout.isCollapsed).to.be.true();
+					expect(await Page.tabLayout.isCollapsed).to.be.true();
 					// go back to tabs
-					$('[data-webos-voice-intent=Select]').moveTo(); // test note: this is different from tabs with icons since this item doesn't exist while expanded, spotlight can't focus it when the mouse moves over the scroller in TabGroup
+					await $('[data-webos-voice-intent=Select]').moveTo(); // test note: this is different from tabs with icons since this item doesn't exist while expanded, spotlight can't focus it when the mouse moves over the scroller in TabGroup
 					// check that layout is not collapsed
-					expect(Page.tabLayout.isCollapsed).to.be.false();
+					expect(await Page.tabLayout.isCollapsed).to.be.false();
 				});
 			});
 		});