+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact sandstone module, newest changes on the top.
 
+## [unreleased]
+
+### Added
+
+- `sandstone/Icon` public class name `icon`
+
 ## [2.1.3] - 2022-03-07
 
 - Updated to use `forwardCustom` and add `type` when forwarding custom events